Database CallBacks and Subject Section Join Commit

+++ app/src/main/java/com/schoolfam/parcher/data/assessmentType/AssessmentType.kt
@@ -5,7 +5,7 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity(tableName = "assessmentTypes")
-data class AssessmentType(@PrimaryKey @ColumnInfo(name = "id") val id: Int,
+data class AssessmentType(@PrimaryKey @ColumnInfo(name = "id") val id: Long,
                      @ColumnInfo(name = "name") val assessment_name: String,
                      @ColumnInfo(name = "maximum_point") val maximumPoint: Double
 )
