+++ app/src/main/java/com/schoolfam/parcher/data/role/Role.kt
@@ -5,6 +5,6 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity(tableName = "roles")
-data class Role( @PrimaryKey @ColumnInfo(name = "id") val id: Int,
+data class Role( @PrimaryKey @ColumnInfo(name = "id") val id: Long,
             @ColumnInfo(name = "name") val name: String
 )
