+++ app/src/main/java/com/schoolfam/parcher/data/section/Section.kt
@@ -5,6 +5,6 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity(tableName = "sections")
-data class Section(@PrimaryKey @ColumnInfo(name = "id") val id: Int,
+data class Section(@PrimaryKey @ColumnInfo(name = "id") val id: Long,
               @ColumnInfo(name = "name") val sectionName: String
 )
